requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
httpx>=0.28,<1.0
python-dotenv>=1.0,<2.0
pywebview>=5.4,<6.0
